ADAS/requirements.txt
# ADAS System Requirements
# Core dependencies for Automated Design of Agentic Systems

# Cerebras API
cerebras-cloud-sdk>=1.0.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Async support
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0

# JSON and data handling
pydantic>=1.8.0

# Logging and monitoring
structlog>=21.1.0

# Optional: PyTorch for advanced model integration (if using local bottleneck models)
# torch>=1.9.0
# torch-geometric>=2.0.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
